Fixes tests and separates the header from the employee table
