+++ src/pages/Employees/components/EmployeeTable/EmployeeTable.test.tsx
@@ -3,16 +3,19 @@ import { render, screen, within, act, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { employees } from "../../../../mocks/fixtures";
 import EmployeeTable from "./EmployeeTable";
+import { MemoryRouter } from "react-router-dom";
 
 describe("EmployeeTable", () => {
   it("has an 'empty' state", async () => {
     render(
-      <EmployeeTable
-        updateEmployee={() => Promise.resolve()}
-        destroyEmployee={(id) => new Promise((resolve) => resolve())}
-        employees={[]}
-        skills={[]}
-      />,
+      <MemoryRouter>
+        <EmployeeTable
+          updateEmployee={() => Promise.resolve()}
+          destroyEmployee={(id) => new Promise((resolve) => resolve())}
+          employees={[]}
+          skills={[]}
+        />,
+      </MemoryRouter>
     );
 
     expect(screen.getByText(/There are currently no team members./i));
@@ -20,12 +23,14 @@ describe("EmployeeTable", () => {
 
   it("shows employees", async () => {
     render(
-      <EmployeeTable
-        updateEmployee={() => Promise.resolve()}
-        destroyEmployee={(id) => new Promise((resolve) => resolve())}
-        employees={employees}
-        skills={[]}
-      />,
+      <MemoryRouter>
+        <EmployeeTable
+          updateEmployee={() => Promise.resolve()}
+          destroyEmployee={(id) => new Promise((resolve) => resolve())}
+          employees={employees}
+          skills={[]}
+        />
+      </MemoryRouter>
     );
 
     employees.forEach((employee) => {
@@ -37,12 +42,14 @@ describe("EmployeeTable", () => {
     const deferred = makeDeferred();
 
     const { findByRole, findAllByRole } = render(
-      <EmployeeTable
-        updateEmployee={() => Promise.resolve()}
-        destroyEmployee={(id) => deferred.promise}
-        employees={employees}
-        skills={[]}
-      />,
+      <MemoryRouter>
+        <EmployeeTable
+          updateEmployee={() => Promise.resolve()}
+          destroyEmployee={(id) => deferred.promise}
+          employees={employees}
+          skills={[]}
+        />,
+      </MemoryRouter>
     );
 
     // click delete icon on the employee table row
