+++ src/pages/Employees/components/EmployeeTable/components/EmployeeTableHeader/index.ts
@@ -1,0 +1,1 @@
+export { default } from './EmployeeTableHeader';
